requirements.txt
certifi>=2022.5.18.1
charset-normalizer>=2.0.12
filehash>=0.2.dev1
idna>=3.3
numpy>=1.22.4
pydantic>=1.9.1
PyYAML>=6.0
requests>=2.28.0
typing_extensions>=4.2.0
urllib3>=1.26.9
paho-mqtt
pylint
